driver/requirements.txt
argparse==1.4.0
pyserial==3.2.1
fs9721==1.2.1
